fix(db): move the persons seed in a separate function and gate with env value
